+++ engine/lib/configuration.php
@@ -655,7 +655,7 @@ function _elgg_load_application_config() {
 	$CONFIG->context = array();
 
 	// needs to be set before system, init for links in html head
-	$CONFIG->lastcache = datalist_get("simplecache_lastupdate");
+	$CONFIG->lastcache = (int)datalist_get("simplecache_lastupdate");
 
 	$CONFIG->i18n_loaded_from_cache = false;
